critnib: add a reference counter and critnib_release()
Add a reference counter and critnib_release() function.

When cb_free_leaf() is SET in critnib_new() the following 4 functions:
- critnib_remove(),
- critnib_get(),
- critnib_find_le() and
- critnib_find()

return a reference (void *ref) to the returned value,
that MUST be released by calling critnib_release()
when it is no longer used and can be freed
using the cb_free_leaf() callback.

Fixes: #1233

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>